Create main.c
Basic Idea was to reduce the number of checks. During writing the algorithm, I realized that
there are a large number of multiple possible paths, from a starting point, but if we look from
the end, there are only three possible paths. So, started calculating the weights from the
endpoints in a backward direction.
For example, let the number of columns be cols, the number of rows be rows, so, if we are in
cols-1 then, for each row, there are three possible points to which we can go from this point. I
calculate the 3 possible sums of weight from this point and then store the minimum weight
among those in the cols-1 of that corresponding row.
After calculating for all rows, I decrement the column and then repeat the same process.
[Basically, they are inside two for loops]. Finally, I end up with rows number of possible
minimum weights if the corresponding row is taken as the starting point. Now I just need to
find the minimum weight from the first column entry of this matrix where I have store the sum
of weight calculations.
Now to calculate the path to be traversed, I maintain another 2D matrix (path_index_entry)
which stores the index of the next hop. I’ll explain with an example.
Considering the above example, so when we are in the cols-1 column we have three possible
row hops in cols column. Let say for minimum weight sum, we take first possible row hop so
I store the row index of that row in the new matrix path_index_entry, similarly, for column
cols-2, there will be three possible hops, I store the row index in the path_index_entry matrix.
Now after populating the matrix completely, when we have the final minimum sum of weight
from the first column from the previous matrix. So now we know the starting point from the
row index of this minimum sum of weight. In the path_index_entry matrix
path_index_entry[row index of this minimum sum of weight][0] will have the entry to the
next hop it needs to take, i.e., path_index_entry[path_index_entry[row index of this
minimum sum of weight][0]][1], and so on.
This is how I solved it.